Refactor grid.lua into separate files now that everything seems to be working.
